io 분리 anonymous socket, chatting socket으로 lib에서 분리(on 해줘야하는 부분이 너무 많아서 namespace로 분리한 것 처럼 폴더도 따로 관리해주기)
