+++ front/src/main.js
@@ -4,7 +4,8 @@ import store from './store'
 import router from './router'
 import vuetify from './plugins/vuetify'
 import VueCookies from "vue-cookies";
-import * as socket from "./lib/socket";
+import * as anonymousSocket from "./lib/anonymousSocket";
+import * as chattingSocket from "./lib/chattingSocket";
 
 
 Vue.config.productionTip = false;
@@ -15,7 +16,7 @@ Vue.use(VueCookies);
 
 
 const init  = async () => {
-  await socket.globalInit();
+  await anonymousSocket.initialize();
   let token = localStorage.getItem('accessToken');
 
   //현재 접속자 수
@@ -29,7 +30,7 @@ const init  = async () => {
   const settingVerifyToken = async () => {
     if(token !== null){
       await store.dispatch('get_user_data', token);
-      await socket.chatInit();
+      await chattingSocket.initialize();
     }
   }
   await settingVerifyToken();
