+++ front/src/store/modules/anonymous.js
@@ -1,7 +1,7 @@
 import jwt_decode from 'jwt-decode'
 import VueCookies from "vue-cookies";
 import anonymous from '../../services/anonymous';
-import * as socket from '../../lib/socket';
+import * as chattingSocket from '../../lib/chattingSocket';
 
 const anonymousModule = {
     state: {
@@ -125,8 +125,8 @@ const anonymousModule = {
                 localStorage.setItem("refreshToken", res.data.refreshToken);
                 await this.dispatch('get_user_data', res.data.token );
                 await this.dispatch('current_user_data'); //소켓서버 접속했음
+                await chattingSocket.initialize();
                 history.back();
-                await socket.chatInit();
                 commit
                 return res.data;
             }else if(res.data.message){
@@ -142,8 +142,9 @@ const anonymousModule = {
                 localStorage.setItem("refreshToken", res.data.refreshToken);       
                 await this.dispatch('get_user_data', res.data.token);
                 await this.dispatch('current_user_data');
+                await chattingSocket.initialize();
                 history.back();
-                await socket.chatInit();
+
                 commit
             }else if(res.data.message){
                 return res.data;
